reorganize, add functional api
